Use a boolean to distinguish an ingredient is in storage instead of a new class
